% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.argument}
\alias{add.argument}
\title{Add an argument to a parser.}
\usage{
add.argument(parser, arg, help, default = NULL, type = NULL, flag = NULL,
  short = NULL)
}
\arguments{
\item{parser}{an \code{arg.parser} object}

\item{arg}{argument name (use no prefix for positional arguments,
\code{--} or \code{-} prefix for optional arguments or flags)}

\item{help}{help description for the argument}

\item{default}{default value for the argument [default: NA]}

\item{type}{variable type of the argument (which can be inferred from
\code{default}), assumed to be \code{character} otherwise}

\item{flag}{whether argument is a flag (and does not consume a value)
[default: FALSE]}

\item{short}{short-form for flags and positional arguments;
short-forms can be assigned automatically based on the first
character of the argument name, unless a conflict arises with
an existing short-form; to avoid conflicts, add the argument
as early as possible}
}
\value{
an \code{arg.parser} object with the argument added
}
\description{
This function is deprecated. Use \code{add_argument} instead.
}

