% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addDirect.R
\docType{methods}
\name{addDirect}
\alias{addDirect}
\alias{addDirect,ArgParser,character-method}
\title{Add directive to parser.}
\usage{
addDirect(x, name, ...)

\S4method{addDirect}{ArgParser,character}(x, name, optional = TRUE,
  help = NULL)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the directive name. Should be of length 1.}

\item{...}{Other arguments used in dispatched method.}

\item{optional}{Optional logical vector of length 1. Is the directive optional?}

\item{help}{Optional character vector shown in usage for the directive. If any, should be of length 1.}
}
\value{
An ArgParser with the directive definition added.
}
\description{
Add directive to parser.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = ArgParser,name = character}: 
}}
\examples{
p <- ArgParser("a test parser")
p <- addDirect(p, "dosomething")
}

