% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFlag.R
\name{addFlag}
\alias{addFlag}
\title{Add flag argument to parser.}
\usage{
addFlag(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the flag name. Should be of length 1 and prefixed with "-{}-".}

\item{...}{Other arguments used in dispatched method.}
}
\description{
Add flag argument to parser.
}
