% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addOpt.R
\docType{methods}
\name{addOpt}
\alias{addOpt}
\alias{addOpt,ArgParser,character-method}
\title{Add positional argument to parser.}
\usage{
addOpt(x, name, ...)

\S4method{addOpt}{ArgParser,character}(x, name, help = NULL, narg = 1L,
  nrequired = narg, directive = NULL)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the opt name. Should be of length 1.}

\item{...}{Other arguments used in dispatched method.}

\item{help}{Optional character vector shown in usage for the opt. If any, should be of length 1.}

\item{narg}{Optinal number of arguments to be consumed.}

\item{nrequired}{Optional number of arguments required to be consumed.}

\item{directive}{character vector indicating any directive the opt belongs to.}
}
\value{
An ArgParser with the opt definition added.
}
\description{
Add positional argument to parser.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = ArgParser,name = character}: 
}}
\examples{
p <- ArgParser("a test parser")
p <- addOpt(p, "filename")
}

