% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSwitch.R
\name{addSwitch}
\alias{addSwitch}
\title{Add switch argument to parser.}
\usage{
addSwitch(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the switch name. Should be of length 1 and prefixed with "-{}-".}

\item{...}{Other arguments used in dispatched method.}
}
\description{
Add switch argument to parser.
}
