% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arg.parser}
\alias{arg.parser}
\title{Create an argument parser.}
\usage{
arg.parser(description, name = NULL)
}
\arguments{
\item{description}{description of the program}

\item{name}{name of the program}
}
\value{
a new arg.parser object
}
\description{
This function creates an arg.parser object. It infers the program name from
the file name of the invoked script.
}
\examples{
p <- arg.parser("A test program");
}

