% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparseR.R
\name{argparse}
\alias{argparse}
\title{Parse arguments of Rscript}
\usage{
argparse(args, flag, default = NULL, type = c("character", "numeric"))
}
\arguments{
\item{args}{vector of characters as from \code{commandArgs(trailingOnly = TRUE)}}

\item{flag}{name of the flag used}

\item{default}{value when flag is not present}

\item{type}{choose type of variable between character (default) and numeric}
}
\value{
returns value of the flag
}
\description{
Parse flags and returns values
}
