% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ArgParser.R
\docType{class}
\name{ArgParser-class}
\alias{ArgParser}
\alias{ArgParser-class}
\title{An S4 class to represent a command line arguments parser.}
\value{
An S4 object of class ArgParser
}
\description{
An S4 class to represent a command line arguments parser.
}
\section{Slots}{

\describe{
\item{\code{desc}}{A description shown in usage. desc with length > 1 will be printed in seperated lines.}

\item{\code{prog}}{A program name shown in usage, if not set, will default to the value of "--file=" flag given in commandArg().}
}}
\examples{
p1 <- ArgParser(desc="a test description", prog="test.R")
p2 <- ArgParser(desc=c("a test description", "that can have multiple lines"))
}

