% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser.R
\docType{package}
\name{argparser}
\alias{argparser}
\title{Command-line argument parser}
\description{
\code{argparser} provides functions for parsing command-line arguments.
}
\details{
To use the parser,
\enumerate{
\item create an \code{arg.parser} object with \code{\link{arg_parser}};
\item add arguments to the parser with \code{\link{add_argument}};
\item call \code{\link{parse_args}} to parse the command line arguments.
}
To execute the script, invoke \code{Rscript}.
Alternatively on Linux, insert a shebang on the first line
(\code{#!/usr/bin/env Rscript}) and \code{chmod +x} the script,
}
