% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{Include R script file}
\usage{
include(file)
}
\arguments{
\item{file}{name}
}
\description{
Include R script with behaviour similar to C++ \code{#include
"header.h"}, by searching in the directory where the 
current script file resides.
}
