% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{parse.args}
\alias{parse.args}
\title{Parse arguments with a parser.}
\usage{
parse.args(parser, argv = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{parser}{an \code{arg.parser} object}

\item{argv}{a character vector to parse (arguments and values should 
already be split by whitespace)}
}
\value{
a list with argument values
}
\description{
This function is deprecated. Use \code{parse_args} instead.
}
