% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCommandLine.R
\name{parseCommandLine}
\alias{parseCommandLine}
\title{Parse command line arguments via given ArgParser.}
\usage{
parseCommandLine(x, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{...}{Other arguments used in dispatched method.}
}
\description{
Parse command line arguments via given ArgParser.
}
