% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.arg.parser}
\alias{print.arg.parser}
\title{Print the help message for an arg.parser.}
\usage{
\method{print}{arg.parser}(x, ...)
}
\arguments{
\item{x}{an arg.parser object}

\item{...}{unused arguments}
}
\description{
This function prints the help message.
}
\details{
At the command line, we would use the \code{--help} or \code{-help} flag
to print the help message:
\code{$ script --help}
}

