% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser.R
\name{show_arg_labels}
\alias{show_arg_labels}
\title{Extract label and help strings from parser.}
\usage{
show_arg_labels(parser)
}
\arguments{
\item{parser}{\code{arg.parser} object}
}
\value{
a list containing a \code{reg.args}, \code{flags}, and
        \code{opt.args} list, which each containing a \code{label}
        string and a \code{help} string
}
\description{
Extract label and help strings from parser.
}
