% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDirect.R
\name{addDirect}
\alias{addDirect}
\title{Add directive to parser.}
\usage{
addDirect(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of directive. Length > 1 is allowed.}

\item{...}{Other arguments used in dispatched method.}
}
\description{
Add directive to parser.
}
