% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addDirect.R
\docType{methods}
\name{addDirect}
\alias{addDirect}
\alias{addDirect,ArgParser,character-method}
\title{Add directive to parser.}
\usage{
addDirect(x, name, ...)

\S4method{addDirect}{ArgParser,character}(x, name, groupname = NULL,
  optional = FALSE, help = NULL)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of directive. Length > 1 is allowed.}

\item{...}{Other arguments used in dispatched method.}

\item{groupname}{Optional character vector of length 1, name for grouped directives. If not supplied, will be auto-named as "_g`i`" for the ith group.}

\item{optional}{Optional logical vector of length 1, indicating wheather the (grouped) directive is optional or not.}

\item{help}{Optional character vector shown in usage for the directive.}
}
\value{
An ArgParser with the directive definition added.
}
\description{
Add directive to parser.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = ArgParser,name = character}: 
}}
\examples{
p <- ArgParser("a test parser")
p <- addDirect(p, "dosomething")
}

