% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addFlag.R
\name{addFlag}
\alias{addFlag}
\title{Add flag argument to parser.}
\usage{
addFlag(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the flag name. Should be of length 1 and prefixed with "-{}-".}

\item{short}{Optional alias for the flag. Should be of length 1 and prefixed with "-".}

\item{default}{Optional default value. If any, Should be of length 1.}

\item{optional}{Optional logical vector of length 1. Is the flag optional?}

\item{help}{Optional character vector shown in usage for the flag. If any, should be of length 1.}
}
\value{
An ArgParser with the flag definition added.
}
\description{
Add flag argument to parser.
}
\examples{
p <- ArgParser("a test parser")
p <- addFlag(p, "--flag", "-f")
}

