% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addOpt.R
\name{addOpt}
\alias{addOpt}
\title{Add positional argument to parser.}
\usage{
addOpt(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the opt name. Should be of length 1.}

\item{help}{Optional character vector shown in usage for the opt. If any, should be of length 1.}

\item{narg}{Optinal number of arguments to be consumed.}

\item{nrequired}{Optional number of arguments required to be consumed.}
}
\value{
An ArgParser with the opt definition added.
}
\description{
Add positional argument to parser.
}
\examples{
p <- ArgParser("a test parser")
p <- addOpt(p, "filename")
}

