% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOpt.R
\name{addOpt}
\alias{addOpt}
\title{Add positional argument to parser.}
\usage{
addOpt(x, name, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{name}{Character vector of the opt name. Should be of length 1.}

\item{...}{Other arguments used in dispatched method.}
}
\description{
Add positional argument to parser.
}
