% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{arg.parser}
\alias{arg.parser}
\title{Create an argument parser.}
\usage{
arg.parser(description, name = NULL)
}
\arguments{
\item{description}{description of the program}

\item{name}{name of the program}
}
\value{
a new \code{arg.parser} object
}
\description{
This function is deprecated. Use \code{arg_parser} instead.
}
