% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arg_parser}
\alias{arg_parser}
\title{Create an argument parser.}
\usage{
arg_parser(description, name = NULL)
}
\arguments{
\item{description}{description of the program}

\item{name}{name of the program}
}
\value{
a new \code{arg.parser} object
}
\description{
This function creates an \code{arg.parser} object. It infers the program
name from the file name of the invoked script.
}
\examples{
p <- arg_parser("A test program")
}

