% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{argparser}
\alias{argparser}
\alias{argparser-package}
\title{Command-line argument parser}
\description{
\code{argparser} provides functions for parsing command-line arguments.
}
\details{
To use the parser,
\enumerate{
\item create an arg.parser object with \code{\link{arg.parser}};
\item add arguments to the parser with \code{\link{add.argument}};
\item call \code{\link{parse.args}} to parse the command line arguments.
}
To execute the script, invoke \code{Rscript}.
Alternatively on Linux, insert a shebang on the first line
(\code{#!/usr/bin/env Rscript}) and \code{chmod +x} the script,
}

