% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseCommandLine.R
\name{parseCommandLine}
\alias{parseCommandLine}
\title{Parse command line arguments via given ArgParser.}
\usage{
parseCommandLine(x, cmdargs, ...)
}
\arguments{
\item{x}{An ArgParser object.}

\item{cmdargs}{A character vector representing command line arguments.}

\item{trim_prefix}{A logical vector indicating wheather to strip dash prefix in argument names in parsed result.}
}
\value{
A list of parsed command line arguments.
}
\description{
Parse command line arguments via given ArgParser.
}
\examples{
library(magrittr)
cmdargs <- strsplit("prog.R --f1 v1", ' ')[[1]]
p <- ArgParser() \%>\%
   addFlag("--f1") \%>\%
   addFlag("--f2")
parseCommandLine(p, cmdargs=cmdargs)
}

